03-call-center/requirements.txt
python-telegram-bot==21.9
groq>=0.9.0
pinecone>=8.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
